Reorganisation de l'équipement en package
