+++ app/src/androidTest/java/cafetcraftteam/rpgsoftware/equipment_test/RangeWeaponTest.java
@@ -1,10 +1,13 @@
-package cafetcraftteam.rpgsoftware;
+package cafetcraftteam.rpgsoftware.equipment_test;
 
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.LargeTest;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import cafetcraftteam.rpgsoftware.equipment.Equipment;
+import cafetcraftteam.rpgsoftware.equipment.RangeWeapon;
 
 import static junit.framework.Assert.assertEquals;
