+++ app/src/main/java/cafetcraftteam/rpgsoftware/equipment/RangeWeapon.java
@@ -1,4 +1,4 @@
-package cafetcraftteam.rpgsoftware;
+package cafetcraftteam.rpgsoftware.equipment;
 
 /**
  * Created by Gautier on 11/02/2016.
